requirements-enhanced.txt
# Enhanced Proposal Processing System Requirements

# Core Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Frontend Dependencies
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0

# Authentication
python-jose[cryptography]>=3.3.0

# AI Integration
google-generativeai>=0.3.0

# Document Processing
python-docx>=0.8.11

# Database
sqlalchemy>=2.0.0

# HTTP Client
requests>=2.31.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: For enhanced features
# Pillow>=10.0.0  # For image processing
# numpy>=1.24.0    # For numerical operations
# python-multipart>=0.0.6  # For file uploads
